Reworked the way we maintain conditions with the system settings (#62)
- We now automatically update the default map to include any active effects from the system effects
- On load, we now automatically update all conditions in the condition map to match the default map unless a condition has been modified by the user
- Removed the system update flow

Overall, these changes are meant to ensure that SUCC remains in line with the base SWADE system without requiring any changes from us or the user. As the status effects are added and existing ones are updated by the SWADE team, SUCC will automatically start using them. For any new effects, it will use the default icon from SWADE and so we'll still have to do a consistency pass at some point, but the functionality will be immediately available to the user without any update from us.